+++ zhihu/common.py
@@ -97,15 +97,15 @@ def check_soup(attr, soup_type='_make_soup'):
     return real
 
 
-def class_common_init(url_re, allowed_none=False):
+def class_common_init(url_re, allowed_none=True):
     def real(func):
         @functools.wraps(func)
         def wrapper(self, url, *args, **kwargs):
             if url is None and not allowed_none:
-                raise ValueError('Invalid Url')
+                raise ValueError('Invalid Url: ' + url)
             if url is not None:
                 if url_re.match(url) is None:
-                    raise ValueError('Invalid URL')
+                    raise ValueError('Invalid URL: ' + url)
                 if url.endswith('/') is False:
                     url += '/'
             if 'session' not in kwargs.keys() or kwargs['session'] is None: